Refactor VGGT readout preparation
