+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "master-thesis-3d-vla"
 version = "0.1.0"
-requires-python = ">=3.10"
+requires-python = ">=3.10,<3.13"
 dependencies = [
     # VGGT deps
     "torch>=2.7",
